requirements.txt
PyPDF2==1.26.0
llama-cpp-python